requirements.txt
tensorflow
tf-keras
transformers
datasets
Pillow
tqdm
pyyaml
numpy